Adding the project
A retrieval system using distributed MapReduce  
I am using bools as documents. Each book undergoes tokenization to remove stop words and stemming to reduce the derived words. 
The distributed MapReduce system then consumes the tokens and produces an inverted index with term-frequency information. The ranking score of the book is computed using different methods (Cosine, Dice and Jaccard) and using TF-IDF weights. The user is presented at the end with the books that ranked highest in similarity using each method (Top ten results).  
The system also includes an interface which prompts the user to enter a search query, and the different jobs that can be run. 
The system is implemented in Scala using Akka actors.